# openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/v1/menus": {"get": {"summary": "Get Menu", "operationId": "get_menu_api_v1_menus_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ShowMenu"}, "type": "array", "title": "Response Get Menu Api V1 Menus Get"}}}}}}, "post": {"summary": "Create Menu", "operationId": "create_menu_api_v1_menus_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/MenuCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShowMenu"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/menus/{target_menu_id}": {"get": {"summary": "Get Uniq Menu", "operationId": "get_uniq_menu_api_v1_menus__target_menu_id__get", "parameters": [{"required": true, "schema": {"type": "string", "title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShowMenu"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Delete Menu", "operationId": "delete_menu_api_v1_menus__target_menu_id__delete", "parameters": [{"required": true, "schema": {"type": "string", "title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Delete Menu Api V1 Menus  Target Menu Id  Delete"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"summary": "Update Menu", "operationId": "update_menu_api_v1_menus__target_menu_id__patch", "parameters": [{"required": true, "schema": {"type": "string", "title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/MenuCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShowMenu"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/menuss": {"get": {"summary": "Get All Data", "operationId": "get_all_data_api_v1_menuss_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/menus/{target_menu_id}/submenus": {"get": {"summary": "Get Submenu", "operationId": "get_submenu_api_v1_menus__target_menu_id__submenus_get", "parameters": [{"required": true, "schema": {"title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"summary": "Create Submenu", "operationId": "create_submenu_api_v1_menus__target_menu_id__submenus_post", "parameters": [{"required": true, "schema": {"title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SubmenuCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShowSubmenu"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/menus/{target_menu_id}/submenus/{target_submenu_id}": {"get": {"summary": "Get Uniq Submenu", "operationId": "get_uniq_submenu_api_v1_menus__target_menu_id__submenus__target_submenu_id__get", "parameters": [{"required": true, "schema": {"title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Submenu Id"}, "name": "target_submenu_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Delete Submenu", "operationId": "delete_submenu_api_v1_menus__target_menu_id__submenus__target_submenu_id__delete", "parameters": [{"required": true, "schema": {"title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Submenu Id"}, "name": "target_submenu_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"summary": "Update Submenu", "operationId": "update_submenu_api_v1_menus__target_menu_id__submenus__target_submenu_id__patch", "parameters": [{"required": true, "schema": {"title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Submenu Id"}, "name": "target_submenu_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SubmenuCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/menus/{target_menu_id}/submenus/{target_submenu_id}/dishes": {"get": {"summary": "Get Dishes", "operationId": "get_dishes_api_v1_menus__target_menu_id__submenus__target_submenu_id__dishes_get", "parameters": [{"required": true, "schema": {"type": "string", "title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}, {"required": true, "schema": {"type": "string", "title": "Target Submenu Id"}, "name": "target_submenu_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ShowDishes"}, "type": "array", "title": "Response Get Dishes Api V1 Menus  Target Menu Id  Submenus  Target Submenu Id  Dishes Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"summary": "Create Dish", "operationId": "create_dish_api_v1_menus__target_menu_id__submenus__target_submenu_id__dishes_post", "parameters": [{"required": true, "schema": {"title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Submenu Id"}, "name": "target_submenu_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Dishescrate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShowDishes"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/menus/{target_menu_id}/submenus/{target_submenu_id}/dishes/{target_dish_id}": {"get": {"summary": "Get Uniq Dish", "operationId": "get_uniq_dish_api_v1_menus__target_menu_id__submenus__target_submenu_id__dishes__target_dish_id__get", "parameters": [{"required": true, "schema": {"title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Submenu Id"}, "name": "target_submenu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Dish Id"}, "name": "target_dish_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShowDishes"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Del Dish", "operationId": "del_dish_api_v1_menus__target_menu_id__submenus__target_submenu_id__dishes__target_dish_id__delete", "parameters": [{"required": true, "schema": {"title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Submenu Id"}, "name": "target_submenu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Dish Id"}, "name": "target_dish_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "title": "Response Del Dish Api V1 Menus  Target Menu Id  Submenus  Target Submenu Id  Dishes  Target Dish Id  Delete"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"summary": "Update Dish", "operationId": "update_dish_api_v1_menus__target_menu_id__submenus__target_submenu_id__dishes__target_dish_id__patch", "parameters": [{"required": true, "schema": {"title": "Target Menu Id"}, "name": "target_menu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Submenu Id"}, "name": "target_submenu_id", "in": "path"}, {"required": true, "schema": {"title": "Target Dish Id"}, "name": "target_dish_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Dishescrate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShowDishes"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Dishescrate": {"properties": {"title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "price": {"type": "string", "title": "Price"}}, "type": "object", "required": ["title", "description", "price"], "title": "Dishescrate"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MenuCreate": {"properties": {"title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}}, "type": "object", "required": ["title", "description"], "title": "MenuCreate"}, "ShowDishes": {"properties": {"orm_mode": {"type": "boolean", "title": "Orm Mode", "default": true}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "price": {"type": "string", "title": "Price"}}, "type": "object", "required": ["id", "title", "description", "price"], "title": "ShowDishes"}, "ShowMenu": {"properties": {"orm_mode": {"type": "boolean", "title": "Orm Mode", "default": true}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "submenus_count": {"type": "integer", "title": "Submenus Count"}, "dishes_count": {"type": "integer", "title": "Dishes Count"}}, "type": "object", "required": ["id", "title", "description", "submenus_count", "dishes_count"], "title": "ShowMenu"}, "ShowSubmenu": {"properties": {"orm_mode": {"type": "boolean", "title": "Orm Mode", "default": true}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "dishes_count": {"type": "integer", "title": "Dishes Count"}}, "type": "object", "required": ["id", "title", "description", "dishes_count"], "title": "ShowSubmenu"}, "SubmenuCreate": {"properties": {"title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}}, "type": "object", "required": ["title", "description"], "title": "SubmenuCreate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}